Rename file name to `amplify-override-slot.directive`
