+++ packages/angular/projects/ui-angular/src/index.ts
@@ -9,7 +9,7 @@ export * from './lib/components';
 export * from './lib/primitives';
 export * from './lib/services/authenticator-context.service';
 export * from './lib/services/state-machine.service';
-export * from './lib/directives/amplify-override.directive';
+export * from './lib/directives/amplify-slot.directive';
 export {
   AmplifyS3Album,
   AmplifyS3ImagePicker,
